+++ packages/causeway-react/src/index.ts
@@ -1,6 +1,7 @@
 export {
   CausewayProvider,
   useCausewayClient,
+  useOptionalCausewayClient,
   useMutation,
   useQuery,
   type CausewayFeedback,