+++ packages/causeway-ts/CHANGELOG.md
@@ -8,6 +8,12 @@ this file. Managed automatically by
 The format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.7-alpha.0](https://github.com/tamimbinhakim/causeway/compare/dyadpy-ts-v0.1.6-alpha.0...causeway-ts-v0.1.7-alpha.0) (2026-05-30)
+
+### Features
+
+- Publish the route-key client transport used by `@causewayjs/client`.
+
 ## [0.1.6-alpha.0](https://github.com/tamimbinhakim/causeway/compare/dyadpy-ts-v0.1.5-alpha.0...dyadpy-ts-v0.1.6-alpha.0) (2026-05-26)